Split shared examples into files
